refactor: extract Signal abstraction out of ValueSignal
This extract the Signal abstraction out of ValueSignal so that the ListSignal implementation can inherit the common functionalities.

Part of #2654
