+++ packages/java/endpoint/src/main/java/com/vaadin/hilla/signals/ValueSignal.java
@@ -3,29 +3,11 @@ package com.vaadin.hilla.signals;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.vaadin.hilla.signals.core.event.StateEvent;
 import jakarta.annotation.Nullable;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Sinks;
 
-import java.util.HashSet;
 import java.util.Objects;
-import java.util.Set;
-import java.util.UUID;
-import java.util.concurrent.locks.ReentrantLock;
 
-public class ValueSignal<T> {
 
-    private static final Logger LOGGER = LoggerFactory
-            .getLogger(ValueSignal.class);
-
-    private final ReentrantLock lock = new ReentrantLock();
-
-    private final UUID id = UUID.randomUUID();
-
-    private final Class<T> valueType;
-
-    private final Set<Sinks.Many<ObjectNode>> subscribers = new HashSet<>();
+public class ValueSignal<T> extends Signal<T> {
 
     private T value;
 
@@ -57,72 +39,7 @@ public class ValueSignal<T> {
      *             <code>null</code>
      */
     public ValueSignal(Class<T> valueType) {
-        Objects.requireNonNull(valueType);
-        this.valueType = valueType;
-    }
-
-    /**
-     * Subscribes to the signal.
-     *
-     * @return a Flux of JSON events
-     */
-    public Flux<ObjectNode> subscribe() {
-        Sinks.Many<ObjectNode> sink = Sinks.many().unicast()
-                .onBackpressureBuffer();
-
-        return sink.asFlux().doOnSubscribe(ignore -> {
-            LOGGER.debug("New Flux subscription...");
-            lock.lock();
-            try {
-                var snapshot = createSnapshotEvent();
-                sink.tryEmitNext(snapshot);
-                subscribers.add(sink);
-            } finally {
-                lock.unlock();
-            }
-        }).doFinally(ignore -> {
-            lock.lock();
-            try {
-                LOGGER.debug("Unsubscribing from Signal...");
-                subscribers.remove(sink);
-            } finally {
-                lock.unlock();
-            }
-        });
-    }
-
-    /**
-     * Submits an event to the signal and notifies subscribers about the change
-     * of the signal value.
-     *
-     * @param event
-     *            the event to submit
-     */
-    public void submit(ObjectNode event) {
-        lock.lock();
-        try {
-            boolean accepted = processEvent(event);
-            // Notify subscribers
-            subscribers.removeIf(sink -> {
-                var eventWithStatus = StateEvent.setAccepted(event, accepted);
-                boolean failure = sink.tryEmitNext(eventWithStatus).isFailure();
-                if (failure) {
-                    LOGGER.debug("Failed push");
-                }
-                return failure;
-            });
-        } finally {
-            lock.unlock();
-        }
-    }
-
-    /**
-     * Returns the signal UUID.
-     *
-     * @return the id
-     */
-    public UUID getId() {
-        return this.id;
+        super(valueType);
     }
 
     /**
@@ -135,7 +52,8 @@ public class ValueSignal<T> {
         return this.value;
     }
 
-    private ObjectNode  createSnapshotEvent() {
+    @Override
+    protected ObjectNode  createSnapshotEvent() {
         var snapshot = new StateEvent<>(getId().toString(),
             StateEvent.EventType.SNAPSHOT, this.value).toJson();
         return StateEvent.setAccepted(snapshot, true);
@@ -153,7 +71,7 @@ public class ValueSignal<T> {
      */
     protected boolean processEvent(ObjectNode event) {
         try {
-            var stateEvent = new StateEvent<>(event, valueType);
+            var stateEvent = new StateEvent<>(event, getValueType());
             return switch (stateEvent.getEventType()) {
                 case SET -> {
                     this.value = stateEvent.getValue();
@@ -188,21 +106,5 @@ public class ValueSignal<T> {
             return true;
         }
         return false;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (!(o instanceof ValueSignal<?> signal)) {
-            return false;
-        }
-        return Objects.equals(getId(), signal.getId());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(getId());
     }
 }
